+++ package.json
@@ -10,6 +10,7 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@react-native-community/datetimepicker": "^6.1.3",
     "@react-native-community/masked-view": "^0.1.11",
     "@react-navigation/bottom-tabs": "^6.3.1",
     "@react-navigation/material-bottom-tabs": "^6.2.1",
@@ -22,6 +23,7 @@
     "react-native-extended-stylesheet": "^0.12.0",
     "react-native-gesture-handler": "^2.4.2",
     "react-native-linear-gradient": "^2.5.6",
+    "react-native-modal-datetime-picker": "^13.1.2",
     "react-native-pager-view": "^5.4.24",
     "react-native-paper": "^4.12.1",
     "react-native-reanimated": "^2.8.0",
